+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -274,6 +274,11 @@ New checks
 
   FIXME: add release notes.
 
+- New :doc:`hooks-state-buf-len
+  <clang-tidy/checks/hooks-state-buf-len>` check.
+
+  FIXME: add release notes.
+
 - New :doc:`hooks-validate-buf-len
   <clang-tidy/checks/hooks-validate-buf-len>` check.